fix: skip certain tests on continuous integration(CI) server
This commit modifies several server test functions to skip their execution in a continuous integration environment. This is achieved by adding a conditional pytest marker to each test, which checks for an environment variable indicating if the tests are being run on the CI server. This change has been made to prevent potential failures on the CI server due to issues specific to its environment.